# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FinTrack API — investment portfolio management and project task tracking
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Portfolio {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  currency: string;
  totalValue: number;
  gainLoss: number;
  gainLossPercent: number;
  createdAt: string;
}

export interface PortfolioInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  currency: string;
}

export interface PortfolioUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
}

export type HoldingAssetType = typeof HoldingAssetType[keyof typeof HoldingAssetType];


export const HoldingAssetType = {
  stock: 'stock',
  etf: 'etf',
  crypto: 'crypto',
  bond: 'bond',
  mutual_fund: 'mutual_fund',
  other: 'other',
} as const;

export interface Holding {
  id: number;
  portfolioId: number;
  ticker: string;
  name: string;
  assetType: HoldingAssetType;
  quantity: number;
  avgCostBasis: number;
  currentPrice: number;
  marketValue: number;
  gainLoss: number;
  gainLossPercent: number;
  createdAt: string;
}

export type HoldingInputAssetType = typeof HoldingInputAssetType[keyof typeof HoldingInputAssetType];


export const HoldingInputAssetType = {
  stock: 'stock',
  etf: 'etf',
  crypto: 'crypto',
  bond: 'bond',
  mutual_fund: 'mutual_fund',
  other: 'other',
} as const;

export interface HoldingInput {
  /** @minLength 1 */
  ticker: string;
  /** @minLength 1 */
  name: string;
  assetType: HoldingInputAssetType;
  quantity: number;
  avgCostBasis: number;
  currentPrice: number;
}

export interface HoldingUpdate {
  quantity?: number;
  avgCostBasis?: number;
  currentPrice?: number;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  buy: 'buy',
  sell: 'sell',
  dividend: 'dividend',
  deposit: 'deposit',
  withdrawal: 'withdrawal',
} as const;

export interface Transaction {
  id: number;
  portfolioId: number;
  ticker: string;
  type: TransactionType;
  quantity: number;
  price: number;
  total: number;
  /** @nullable */
  notes?: string | null;
  transactedAt: string;
  createdAt: string;
}

export type TransactionInputType = typeof TransactionInputType[keyof typeof TransactionInputType];


export const TransactionInputType = {
  buy: 'buy',
  sell: 'sell',
  dividend: 'dividend',
  deposit: 'deposit',
  withdrawal: 'withdrawal',
} as const;

export interface TransactionInput {
  /** @minLength 1 */
  ticker: string;
  type: TransactionInputType;
  quantity: number;
  price: number;
  notes?: string;
  transactedAt: string;
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  active: 'active',
  on_hold: 'on_hold',
  completed: 'completed',
  archived: 'archived',
} as const;

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  status: ProjectStatus;
  /** @nullable */
  color?: string | null;
  progress: number;
  taskCount: number;
  completedCount: number;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
}

export type ProjectInputStatus = typeof ProjectInputStatus[keyof typeof ProjectInputStatus];


export const ProjectInputStatus = {
  active: 'active',
  on_hold: 'on_hold',
  completed: 'completed',
  archived: 'archived',
} as const;

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  status?: ProjectInputStatus;
  color?: string;
  dueDate?: string;
}

export type ProjectUpdateStatus = typeof ProjectUpdateStatus[keyof typeof ProjectUpdateStatus];


export const ProjectUpdateStatus = {
  active: 'active',
  on_hold: 'on_hold',
  completed: 'completed',
  archived: 'archived',
} as const;

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  status?: ProjectUpdateStatus;
  color?: string;
  dueDate?: string;
}

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  review: 'review',
  done: 'done',
} as const;

export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];


export const TaskPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  urgent: 'urgent',
} as const;

export interface Task {
  id: number;
  projectId: number;
  /** @nullable */
  projectName?: string | null;
  title: string;
  /** @nullable */
  description?: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  /** @nullable */
  ownerId?: number | null;
  /** @nullable */
  ownerName?: string | null;
  /** @nullable */
  ownerInitials?: string | null;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
}

export type TaskInputStatus = typeof TaskInputStatus[keyof typeof TaskInputStatus];


export const TaskInputStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  review: 'review',
  done: 'done',
} as const;

export type TaskInputPriority = typeof TaskInputPriority[keyof typeof TaskInputPriority];


export const TaskInputPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  urgent: 'urgent',
} as const;

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  status: TaskInputStatus;
  priority: TaskInputPriority;
  ownerId?: number;
  dueDate?: string;
}

export type TaskUpdateStatus = typeof TaskUpdateStatus[keyof typeof TaskUpdateStatus];


export const TaskUpdateStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  review: 'review',
  done: 'done',
} as const;

export type TaskUpdatePriority = typeof TaskUpdatePriority[keyof typeof TaskUpdatePriority];


export const TaskUpdatePriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  urgent: 'urgent',
} as const;

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  status?: TaskUpdateStatus;
  priority?: TaskUpdatePriority;
  /** @nullable */
  ownerId?: number | null;
  /** @nullable */
  dueDate?: string | null;
}

export interface Member {
  id: number;
  name: string;
  initials: string;
  role: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  avatarColor?: string | null;
  createdAt: string;
}

export interface MemberInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  role: string;
  email?: string;
  avatarColor?: string;
}

export interface MemberUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  role?: string;
  email?: string;
  avatarColor?: string;
}

export interface DashboardSummary {
  totalPortfolioValue: number;
  totalGainLoss: number;
  totalGainLossPercent: number;
  portfolioCount: number;
  totalTasks: number;
  openTasks: number;
  overdueTasks: number;
  projectCount: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  transaction: 'transaction',
  task_created: 'task_created',
  task_updated: 'task_updated',
  task_completed: 'task_completed',
  project_created: 'project_created',
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  description: string;
  /** @nullable */
  meta?: string | null;
  timestamp: string;
}

export interface BreakdownItem {
  label: string;
  count: number;
}

export interface TaskBreakdown {
  byStatus: BreakdownItem[];
  byPriority: BreakdownItem[];
}