# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FinTrack API — investment portfolio management and project task tracking
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /portfolios:
    get:
      operationId: listPortfolios
      tags: [portfolios]
      summary: List all portfolios
      responses:
        "200":
          description: List of portfolios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Portfolio"
    post:
      operationId: createPortfolio
      tags: [portfolios]
      summary: Create a portfolio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
  /portfolios/{id}:
    get:
      operationId: getPortfolio
      tags: [portfolios]
      summary: Get a portfolio
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Portfolio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
        "404":
          description: Not found
    patch:
      operationId: updatePortfolio
      tags: [portfolios]
      summary: Update a portfolio
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
    delete:
      operationId: deletePortfolio
      tags: [portfolios]
      summary: Delete a portfolio
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Holdings ────────────────────────────────────────────────────────────────
  /portfolios/{portfolioId}/holdings:
    get:
      operationId: listHoldings
      tags: [holdings]
      summary: List holdings in a portfolio
      parameters:
        - name: portfolioId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Holdings list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Holding"
    post:
      operationId: createHolding
      tags: [holdings]
      summary: Add a holding to a portfolio
      parameters:
        - name: portfolioId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HoldingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Holding"
  /holdings/{id}:
    patch:
      operationId: updateHolding
      tags: [holdings]
      summary: Update a holding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HoldingUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Holding"
    delete:
      operationId: deleteHolding
      tags: [holdings]
      summary: Delete a holding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Transactions ─────────────────────────────────────────────────────────
...
  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List all team members
      responses:
        "200":
          description: Members list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      operationId: createMember
      tags: [members]
      summary: Create a team member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{id}:
    patch:
      operationId: updateMember
      tags: [members]
      summary: Update a team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
    delete:
      operationId: deleteMember
      tags: [members]
      summary: Delete a team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Dashboard ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Portfolio:
      type: object
      required: [id, name, currency, totalValue, gainLoss, gainLossPercent, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        currency:
          type: string
        totalValue:
          type: number
        gainLoss:
          type: number
        gainLossPercent:
          type: number
        createdAt:
          type: string
    PortfolioInput:
      type: object
      required: [name, currency]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        currency:
          type: string
    PortfolioUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string

    # ── Holding schemas ─────────────────────────────────────────────────────
    Holding:
      type: object
      required: [id, portfolioId, ticker, name, assetType, quantity, avgCostBasis, currentPrice, marketValue, gainLoss, gainLossPercent, createdAt]
      properties:
        id:
          type: integer
        portfolioId:
          type: integer
        ticker:
          type: string
        name:
          type: string
        assetType:
          type: string
          enum: [stock, etf, crypto, bond, mutual_fund, other]
        quantity:
          type: number
        avgCostBasis:
          type: number
        currentPrice:
          type: number
        marketValue:
          type: number
        gainLoss:
          type: number
        gainLossPercent:
          type: number
        createdAt:
          type: string
    HoldingInput:
      type: object
      required: [ticker, name, assetType, quantity, avgCostBasis, currentPrice]
      properties:
        ticker:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        assetType:
          type: string
          enum: [stock, etf, crypto, bond, mutual_fund, other]
        quantity:
          type: number
        avgCostBasis:
          type: number
        currentPrice:
          type: number
    HoldingUpdate:
      type: object
      properties:
        quantity:
          type: number
        avgCostBasis:
          type: number
        currentPrice:
          type: number

    # ── Transaction schemas ─────────────────────────────────────────────────
...
    Member:
      type: object
      required: [id, name, initials, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        initials:
          type: string
        role:
          type: string
        email:
          type: ["string", "null"]
        avatarColor:
          type: ["string", "null"]
        createdAt:
          type: string
    MemberInput:
      type: object
      required: [name, role]
      properties:
        name:
          type: string
          minLength: 1
        role:
          type: string
          minLength: 1
        email:
          type: string
        avatarColor:
          type: string
    MemberUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        role:
          type: string
          minLength: 1
        email:
          type: string
        avatarColor:
          type: string

    # ── Dashboard schemas ───────────────────────────────────────────────────
...