# lib/api-zod/src/generated/types/holding.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FinTrack API — investment portfolio management and project task tracking
 * OpenAPI spec version: 0.1.0
 */
import type { HoldingAssetType } from './holdingAssetType';

export interface Holding {
  id: number;
  portfolioId: number;
  ticker: string;
  name: string;
  assetType: HoldingAssetType;
  quantity: number;
  avgCostBasis: number;
  currentPrice: number;
  marketValue: number;
  gainLoss: number;
  gainLossPercent: number;
  createdAt: string;
}