# lib/api-zod/src/generated/types/holdingUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FinTrack API — investment portfolio management and project task tracking
 * OpenAPI spec version: 0.1.0
 */

export interface HoldingUpdate {
  quantity?: number;
  avgCostBasis?: number;
  currentPrice?: number;
}