# lib/api-zod/src/generated/types/portfolio.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FinTrack API — investment portfolio management and project task tracking
 * OpenAPI spec version: 0.1.0
 */

export interface Portfolio {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  currency: string;
  totalValue: number;
  gainLoss: number;
  gainLossPercent: number;
  createdAt: string;
}